requirements.txt
streamlit
google-genai
matplotlib